router structure, css improvements
