+++ my-react-app/src/App.tsx
@@ -1,7 +1,8 @@
+import router from "./Router/Router";
 import "./App.css";
 
 import { RouterProvider } from "react-router-dom";
-import router from "./Router/Router";
+
 
 function App() {
 	return <RouterProvider router={router} />;
