+++ my-react-app/src/Pages/ButtonHome/ButtonHome.css
@@ -1,0 +1,32 @@
+
+.circle {
+    width: 300px;
+    height: 300px;
+    background-color: var(--main-color-red);
+    border-radius: 50%;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    font-size: 2rem;
+    color: var(--main-color-text-light);
+    font-family: var(--button-font-family);
+    cursor: pointer;
+    transition: all 0.3s;
+    text-decoration: none;
+    box-shadow: 7px 7px 10px gray      
+}
+
+.text {
+    font-size: var(--titre-font-size);
+    font-family: var(--main-titre-font-family);
+    color: var(--main-color-text-light);
+    margin: var(--main-margin);
+    font-weight: bold;
+    text-decoration: none;
+}    
+
+
+.circle:active {
+    transform: translateY(0%);
+    box-shadow: inset 5px 10px 15px black;
+}
