--- my-react-app/src/Pages/FetchPoem.css
@@ -1,14 +1,0 @@
-.container-poem {
-    width: 100%;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    
-}
-
-.poem {
-    white-space: pre-wrap;
-    word-wrap: break-word; 
-    overflow-wrap: break-word; 
-    text-align: left; 
-  }
