+++ my-react-app/src/Pages/SectionMeme/SectionMeme.css
@@ -1,0 +1,18 @@
+.meme {
+    width: 500px;
+    height: fit-content;
+    padding: 40px 40px 40px 40px;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+}
+
+.container_buttons {
+    justify-content: space-between;
+    align-items: center;
+    margin-top: 20px;
+    width: 100%;
+    max-width: 500px;
+    margin: 20px auto;
+
+}
