+++ my-react-app/src/Pages/SectionMeme/SectionMeme.tsx
@@ -2,8 +2,9 @@ import React from 'react';
 import { useState, useEffect } from 'react';
 import axios from 'axios';
 import './SectionMeme.css';
-import ButtonRecharge from '../ButtonRecharge/ButtonRecharge';
-import ButtonRetour from '../ButtonRetour/ButtonRetour';
+import ButtonRecharge from '../../components/ButtonRecharge/ButtonRecharge';
+import ButtonRetour from '../../components/ButtonRetour/ButtonRetour';
+import Panneau from '../../components/Panneau/Panneau';
 
 export default function SectionMeme () {
 const [meme, setMeme] = useState('https://static.demilked.com/wp-content/uploads/2024/09/random-memes-funny-5.jpeg');
@@ -27,7 +28,7 @@ const getMeme = async () => {
 
     return (
             <div className = "pageMeme">
-                <div className = "container">
+            <Panneau>
         {meme ? (
         <img
             src={meme}
@@ -38,10 +39,10 @@ const getMeme = async () => {
         <p>Chargement du mème...</p>
         )}
 
-        </div>
+</Panneau>
                 <div className = "container_buttons">
                 <ButtonRecharge onClick={getMeme} className = ".button-recharge" />
-                <ButtonRetour onClick={getMeme} className = ".button-retour" />
+                <ButtonRetour />
                 </div>
         </div>
