+++ my-react-app/src/Router/Router.tsx
@@ -1,18 +1,33 @@
 import { createBrowserRouter } from "react-router-dom";
 
 
+import ButtonHome from "../Pages/ButtonHome/ButtonHome.tsx";
 import ThreeButtonsPage from "../components/ThreeButtonsPage/ThreeButtons.tsx";
-import DisplayPoem from "../pages/DisplayPoem.tsx";
+import SectionPoem from "../Pages/SectionPoem/SectionPoem.tsx";
+import SectionMeme from "../Pages/SectionMeme/SectionMeme.tsx";
+import SectionLearn from "../Pages/SectionLearn/SectionLearn.tsx";
 
 const router = createBrowserRouter([
 	{
 		path: "/",
+		element: <ButtonHome />,
+	},
+	{
+		path: "/home",
 		element: <ThreeButtonsPage />,
 	},
 	{
-		path: "/inspire",
-		element: <DisplayPoem />,
+		path: "/laugh",
+		element: <SectionMeme />,
 	},
+	{
+		path: "/inspire",
+		element: <SectionPoem />,
+	},
+	{
+		path: "/learn",
+		element: <SectionLearn />,
+	}
 ]);
 
 export default router;
