--- my-react-app/src/components/ButtonHome/ButtonHome.css
@@ -1,69 +1,0 @@
-.container {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    height: 100vh;
-    margin: 0;
-    position: relative;
-    z-index: 1; 
-      
-      
-}
-
-.circle {
-    width: 300px;
-    height: 300px;
-    background-color: var(--main-color-red);
-    border-radius: 50%;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    font-size: 2rem;
-    color: var(--main-color-text-light);
-    font-family: var(--button-font-family);
-    cursor: pointer;
-    transition: all 0.3s;
-    box-shadow: 7px 7px 10px gray      
-}
-
-.text {
-    font-size: var(--titre-font-size);
-    font-family: var(--main-titre-font-family);
-    color: var(--main-color-text-light);
-    margin: var(--main-margin);
-    font-weight: bold;
-}    
-  
-.button-home-container {
-   position: relative;
-   display: flex;
-   justify-content: center;
-   align-items: center;
-   height: 100vh; 
-   background-color: #f0f0f0;
-}
-
-.button-home-container .panneau {
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-    z-index: 0; 
-}  
-  
-.panneau {
-    position: absolute; 
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-    width: 300px; 
-    height: 300px; 
-    background-color: #FBFAFA;
-    border-radius: 5px;
-    z-index: 0; 
-} 
-  
-.circle:active {
-    transform: translateY(0%);
-    box-shadow: inset 5px 10px 15px black;
-}
