+++ my-react-app/src/components/ButtonRetour/ButtonRetour.tsx
@@ -1,12 +1,15 @@
 import React from 'react';
+import { Link } from "react-router-dom";
 import './ButtonRetour.css';
 import { useState, useEffect } from 'react';
 // import ThreeButtons from '../ThreeButtonsPage/ThreeButtons';
 
-export default function ButtonRetour({ onClick }) {
-    return (
-    <div>
-        <button className="button-retour" onClick={()=> onClick}><img src='../assets/icon_home.png'/></button>
-    </div>
-    );
-};
+export default function ButtonRetour() {
+	return (
+		<div>
+			<Link to="/home" className="button-retour">
+				<img src="../assets/icon_home.png" alt="Home" />
+			</Link>
+		</div>
+	);
+}
