+++ my-react-app/src/components/PaginationControls/PaginationControls.css
@@ -15,10 +15,10 @@ button {
   }
 
   
-  div {
+  /* div {
     display: flex;
     justify-content: center; 
     gap: 20px; 
     margin-top: 20px; 
   }
-  
+   */
