+++ my-react-app/src/components/Panneau/Panneau.css
@@ -1,13 +1,3 @@
-
-.fond{    
-    display: flex;
-    justify-content: center; 
-    align-items: center;             
-    margin: 0;              
-    background-color: #C8C6BF;
-    padding: 1rem 0;
-}
-
 .panneau{           
     box-sizing: border-box;   
     background-color: #FBFAFA;           
@@ -18,6 +8,7 @@
     height: 95vh;
     overflow-y: scroll;
     padding: 3rem 1.5rem;
+    margin: 2rem;
 }
 
 .panneau .vis1,
