--- my-react-app/src/components/SectionMeme/SectionMeme.css
@@ -1,30 +1,0 @@
-.pageMeme {
-    height: 100vh;
-    background: var(--main-color-amuser)
-}
-
-.container {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    height: 70vh;
-    background-color: #D9D9D9;
-    margin: 0;
-    padding: 40px 40px 40px 80px;
-}
-
-.meme {
-    width: 500px;
-    height: fit-content;
-    padding: 40px 40px 40px 40px;
-}
-
-.container_buttons {
-    justify-content: space-between;
-    align-items: center;
-    margin-top: 20px;
-    width: 100%;
-    max-width: 500px;
-    margin: 20px auto;
-
-}
